pyproject.toml
[tool.poetry]
name = "remindme"
version = "0.1.0"
description = ""
authors = ["Matt Nicholson <matthewn1991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
